Qt: Guard better on profile ZIP archives
The Qt documentation has to be read very carefully to see which
functions remove ".." and which don't. Add more explicit comments
noting that absolute paths are allowed on Windows because of an
error creating profile ZIP archives up until 3.2.5 (but maybe we
should stop as that's a no longer supported version.) After converting
absolute paths to relative, perform a directory traversal check on
all OSes.

Also, don't cast the returned uncompressed_size in the unz_file_info64
struct from a ZPOS64_T (uint64_t) to an int, to handle the case where
a ZIP64 archive has a file with a 64-bit file size that when cast to
a 32-bit signed integer becomes negative or otherwise less than 256 MiB.

Fixup 70c86bd39c74fc0052c9a212afc911b41d612388

Fix #21553

(backported from commit 683ee3badeebecbf4409b76d649ff43bd54e368b)

+++ ui/qt/models/profile_model.h
@@ -142,7 +142,19 @@ private:
     int findAsReference(QString reference) const;
 
 #if defined(HAVE_MINIZIP) || defined(HAVE_MINIZIPNG)
-    static bool acceptFile(QString fileName, int fileSize);
+    /**
+     * @brief Decides whether a file from a ZIP archive should be included in the import.
+     * @param fileName Name of the file entry within the archive.
+     * @param fileSize Uncompressed size of the file in bytes.
+     * @return @c true if the file should be extracted and imported.
+     */
+    static bool acceptFile(QString fileName, uint64_t fileSize);
+
+    /**
+     * @brief Sanitises a filename extracted from a ZIP archive for use as a profile name.
+     * @param fileName Raw filename from the archive.
+     * @return Sanitised name safe for use as a profile directory name.
+     */
     static QString cleanName(QString fileName);
 #endif
